+++ server/internal/archive/dimag/soap_models.go
@@ -4,8 +4,39 @@ import (
 	"encoding/xml"
 )
 
-const soapNs string = "http://schemas.xmlsoap.org/soap/envelope/"
-const dimagNs string = "http://dimag.la-bw.de/WebService.wsdl"
+// General
+
+type envelope[T any] struct {
+	XMLName xml.Name `xml:"soapenv:Envelope"`
+	SoapNs  string   `xml:"xmlns:soapenv,attr"`
+	DimagNs string   `xml:"xmlns:web,attr"`
+	Header  struct {
+		XMLName xml.Name `xml:"soapenv:Header"`
+	}
+	Body struct {
+		XMLName xml.Name `xml:"soapenv:Body"`
+		Data    T
+	}
+}
+
+func makeEnvelope[T any](data T) envelope[T] {
+	envelope := envelope[T]{
+		SoapNs:  "http://schemas.xmlsoap.org/soap/envelope/",
+		DimagNs: "http://dimag.la-bw.de/WebService.wsdl",
+	}
+	envelope.Body.Data = data
+	return envelope
+}
+
+type responseEnvelope[T any] struct {
+	XMLName xml.Name `xml:"Envelope"`
+	SoapNs  string   `xml:"xmlns:SOAP-ENV,attr"`
+	DimagNs string   `xml:"xmlns:ns1,attr"`
+	Body    struct {
+		XMLName xml.Name `xml:"Body"`
+		Data    T
+	}
+}
 
 // ImportBag
 
@@ -14,28 +45,7 @@ type importBagData struct {
 	UserName  string   `xml:"username"`
 	Password  string   `xml:"password"`
 	BagItPath string   `xml:"bagItPath"`
-}
-
-type importBagEnvelope struct {
-	XMLName xml.Name `xml:"soapenv:Envelope"`
-	SoapNs  string   `xml:"xmlns:soapenv,attr"`
-	DimagNs string   `xml:"xmlns:web,attr"`
-	Header  struct {
-		XMLName xml.Name `xml:"soapenv:Header"`
-	}
-	Body struct {
-		XMLName   xml.Name `xml:"soapenv:Body"`
-		ImportBag importBagData
-	}
-}
-
-func makeImportBagEnvelope(data importBagData) importBagEnvelope {
-	envelope := importBagEnvelope{
-		SoapNs:  soapNs,
-		DimagNs: dimagNs,
-	}
-	envelope.Body.ImportBag = data
-	return envelope
+	Async     bool     `xml:"async"`
 }
 
 type importBagResponse struct {
@@ -47,39 +57,25 @@ type importBagResponse struct {
 	Types   string   `xml:"types"`
 }
 
-type importBagResponseEnvelope struct {
-	XMLName xml.Name `xml:"Envelope"`
-	SoapNs  string   `xml:"xmlns:SOAP-ENV,attr"`
-	DimagNs string   `xml:"xmlns:ns1,attr"`
-	Body    struct {
-		XMLName           xml.Name `xml:"Body"`
-		ImportBagResponse importBagResponse
-	}
+// GetJobStatus
+
+type getJobStatusData struct {
+	XMLName  xml.Name `xml:"web:getJobStatus"`
+	UserName string   `xml:"username"`
+	Password string   `xml:"password"`
+	JobID    int      `xml:"jid"`
+}
+
+type getJobStatusResponse struct {
+	XMLName xml.Name `xml:"getJobStatusResponse"`
+	Status  int      `xml:"status"`
+	JobID   int      `xml:"jid"`
+	Message string   `xml:"msg"`
+	AIDs    string   `xml:"aids"`
+	Types   string   `xml:"types"`
 }
 
 // GetAIDforKeyValue
-
-type getAIDForKeyValueEnvelope struct {
-	XMLName xml.Name `xml:"soapenv:Envelope"`
-	SoapNs  string   `xml:"xmlns:soapenv,attr"`
-	DimagNs string   `xml:"xmlns:web,attr"`
-	Header  struct {
-		XMLName xml.Name `xml:"soapenv:Header"`
-	}
-	Body struct {
-		XMLName           xml.Name `xml:"soapenv:Body"`
-		GetAIDforKeyValue getAIDForKeyValueData
-	}
-}
-
-func makeGetAIDForKeyValueEnvelope(data getAIDForKeyValueData) getAIDForKeyValueEnvelope {
-	envelope := getAIDForKeyValueEnvelope{
-		SoapNs:  soapNs,
-		DimagNs: dimagNs,
-	}
-	envelope.Body.GetAIDforKeyValue = data
-	return envelope
-}
 
 type getAIDForKeyValueData struct {
 	XMLName  xml.Name `xml:"web:getAIDforKeyValue"`
@@ -88,16 +84,6 @@ type getAIDForKeyValueData struct {
 	Key      string   `xml:"key"`
 	Value    string   `xml:"value"`
 	Typ      string   `xml:"typ"`
-}
-
-type getAIDForKeyValueResponseEnvelope struct {
-	XMLName xml.Name `xml:"Envelope"`
-	SoapNs  string   `xml:"xmlns:SOAP-ENV,attr"`
-	DimagNs string   `xml:"xmlns:ns1,attr"`
-	Body    struct {
-		XMLName                   xml.Name `xml:"Body"`
-		GetAIDforKeyValueResponse getAIDForKeyValueResponse
-	}
 }
 
 type getAIDForKeyValueResponse struct {